Camera API 2

+++ TimelapserAndroidApp/app/src/main/java/com/courcelle/timelapser/utils/FileUtils.java
@@ -3,13 +3,16 @@ package com.courcelle.timelapser.utils;
 import java.io.File;
 
 public class FileUtils {
-    public static int remoteOldFiles(File path,long olderThan) {
+    public static int removeOldFiles(File path, long olderThan) {
         int count=0;
         if (path.isDirectory()) {
-            for(File file: path.listFiles()) {
-                if (file.lastModified()<olderThan) {
-                    if (file.delete()) {
-                        count++;
+            File[] files = path.listFiles();
+            if (files!=null) {
+                for(File file: files) {
+                    if (file.lastModified()<olderThan) {
+                        if (file.delete()) {
+                            count++;
+                        }
                     }
                 }
             }
